+++ frontend/app/src/components/inputs/AmountInput.vue
@@ -83,19 +83,10 @@ defineExpose({
   >
     <!-- Pass on all named slots -->
     <slot v-for="slot in Object.keys(slots)" :slot="slot" :name="slot" />
+
     <!-- Pass on all scoped slots -->
-    <template
-      v-for="slot in Object.keys($scopedSlots)"
-      :slot="slot"
-      slot-scope="scope"
-    >
-      <slot
-        :name="slot"
-        v-bind="
-          // @ts-ignore
-          scope
-        "
-      />
+    <template v-for="slot in Object.keys($scopedSlots)" #[slot]="scope">
+      <slot v-bind="scope" :name="slot" />
     </template>
   </v-text-field>
 </template>